Create synatx_error.js
The error occurs due to the incorrect function call calculate Sum (numbers); there should not be a space between calculate and Sum. In JavaScript, function names cannot have spaces.

JavaScript Error Handling:

JavaScript throws a SyntaxError, providing the message "Unexpected identifier", which means the interpreter encountered a part of the code that does not conform to the expected syntax (in this case, the space between calculate and Sum).